requirements.txt
torch
scikit-learn
pandas
numpy
biopython
torchtext